requirements.txt
Flask--3.0.0
pandas--2.0.3
numpy--1.24.4
requests--2.31.0
pickle--3.11.0
gunicorn--21.2.0
scikit-learn--1.3.2